% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_s.R
\name{mk_s}
\alias{mk_s}
\title{Mann Kendall S statistic}
\usage{
mk_s(x)
}
\arguments{
\item{x}{time series}
}
\value{
Mann-Kendall's S statistic
}
\description{
Calculates the MK S statistic. Internal function used in mk_z_stat.
}
