% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_z_stat.R
\name{mk_z_stat}
\alias{mk_z_stat}
\title{Mann Kendall Trend test, corrected for AR(1)}
\usage{
mk_z_stat(x)
}
\arguments{
\item{x}{time series of a single grid cell}
}
\value{
Mann-Kendall's S statistic transformed to a Z score
}
\description{
Computes the Mann-Kendall trend test, S, and converts it into a Z score.
}
