% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs-methods.R
\name{nobs-methods}
\alias{nobs-methods}
\alias{nobs.numeric}
\alias{nobs.integer}
\alias{nobs.matrix}
\alias{nobs.data.frame}
\alias{nobs.character}
\alias{nobs.factor}
\title{Number of observations in a given object}
\usage{
\method{nobs}{numeric}(object, ...)

\method{nobs}{integer}(object, ...)

\method{nobs}{matrix}(object, ...)

\method{nobs}{data.frame}(object, ...)

\method{nobs}{character}(object, ...)

\method{nobs}{factor}(object, ...)
}
\arguments{
\item{object}{A data frame or matrix, or a numeric, integer, character, or
factor vector.}

\item{...}{Arguments passed to other methods.}
}
\value{
The numeric number of observations in \code{object}.
}
\description{
\code{\link[stats:nobs]{stats::nobs()}} is a generic function that returns the number of
observations from a model. permute provides methods for several other types
of R object.
}
\details{
These methods return the number of observations in numeric, integer,
character, or factor vectors, matrices, and data frames.
}
\examples{
\dontshow{suppressWarnings(RNGversion("3.5.0"))}
set.seed(1)
## numeric vector
len <- sample(1:10, 1)
v <- as.numeric(sample(1:100, len))
len
obs <- nobs(v)
isTRUE(all.equal(len, obs))

## integer
len <- sample(1L:10L, 1)
obs <- nobs(len)
isTRUE(all.equal(len, obs))
}
\author{
Gavin Simpson
}
