% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_dist.R
\name{perm_dist}
\alias{perm_dist}
\title{Permutation distribution}
\usage{
perm_dist(
  data,
  fx,
  nperm = 1000,
  alpha_local,
  alpha_global,
  null_distribution,
  block_size = NULL,
  seed,
  verbose = TRUE
)
}
\arguments{
\item{data}{Data in a 3d array, where the first two dimensions are the
physical dimensions (e.g., lon and lat), and the third one is time.}

\item{fx}{function to be applied at each grid cell. Should be self
sufficient - no extra arguments needed. Should return only the test
statistic}

\item{nperm}{number of permutations. Defaults to 1000}

\item{verbose}{Counter returning when the function is done with 10 function
calls}
}
\value{
returns the distribution of maxT and stcs, each a vector in a list
}
\description{
Obtains the permutation distribution of the maximum statistic and the STCS,
named maxT and stcs, respectively. Returns a list containing maxT and stcs:
each a vector of length nperm where each entry is the maximum of the nth
permutation. Together form the permutation distribution, of which
the (1-alpha)th percentile is the threshold of significance adjusted
for multiple testing.
}
