% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_matrix.R
\name{perm_matrix}
\alias{perm_matrix}
\title{Permutation Matrix}
\usage{
perm_matrix(nobs, nperm, block_size = NULL, seed = NULL)
}
\arguments{
\item{nobs}{number of observations to permute}

\item{nperm}{number of permutations, including original data ordering.}

\item{block_size}{Desired block size for block permutation.}

\item{seed}{seed to be fed into set.seed function}
}
\value{
matrix for use in permutation methods. Each row is one permutation
}
\description{
Generates matrix for permutaitons. Each row is a permutation. Makes it easy
to do block permutations. Repeated lines are avoided. Original ordering
is always the last row.
}
