\name{permute-deprecated}

\title{Deprecated functions in package permute}

\description{
  These functions have been replaced in the latest version of
  permute. They are provided in the package to allow time for the
  transition of code to the replacements in the package. The functions
  listed here will soon be removed from the package, possibly by the
  next minor (x.y) release.
}

\details{
  \code{permCheck} was deprecated in favour of function
  \code{\link{check}}. The documentation for \code{permCheck} can be
  found via \code{help("permCheck-deprecated")} (note the quotes).

  \code{permControl} was deprecated in favour of of function
  \code{\link{how}}. The documentation for \code{permControl} can be
  found via \code{help("permControl-deprecated")} (note the quotes).
}

\author{Gavin L. Simpson}

\seealso{\code{\link{check}}, and \code{\link{how}}.}