% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mtc.R
\name{print.mtc}
\alias{print.mtc}
\title{Print function for mtc class}
\usage{
\method{print}{mtc}(object)
}
\arguments{
\item{object}{an object of class mtr}
}
\description{
Print the summary of the test statistic, p values, and detected grid cells
for each mutliple testing correction.
}
