% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk_fn.R
\name{rk_fn}
\alias{rk_fn}
\title{Lag k serial correlation}
\usage{
rk_fn(x, k = 1)
}
\arguments{
\item{x}{time series}
}
\value{
the lag-k serial correlation coefficient
}
\description{
computes lag-k serial correlation coefficient of vector x , r_k
}
