% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mtc.R
\name{summary.mtc}
\alias{summary.mtc}
\title{Summary function for mtc class}
\usage{
\method{summary}{mtc}(object)
}
\arguments{
\item{object}{an object of class mtr}
}
\description{
Print the summary of the test statistic, p values, and detected grid cells
for each mutliple testing correction.
}
