% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_gistemp.R
\docType{data}
\name{temp_gistemp}
\alias{temp_gistemp}
\title{Gridded world temperature data .}
\format{
An array with dimensions 180, 90, 68
\describe{
  \item{lon}{longitude, in degrees}
  \item{lat}{latitude, in degrees}
  \item{t}{time, in years, starting in 1951}
}
}
\source{
\url{https://data.giss.nasa.gov/pub/gistemp/GHCNv3/gistemp1200_ERSSTv5.nc.gz}
}
\usage{
temp_gistemp
}
\description{
A dataset containing the temperature at each grid cell in the world.
}
\keyword{datasets}
