% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_data.R
\name{threshold_data}
\alias{threshold_data}
\title{Threshold  data
Thresholds the data according to the corresponding multiple testing
correction. This function is used internally.}
\usage{
threshold_data(
  perm_results,
  alpha_local,
  alpha_global,
  data_dim,
  null_distribution
)
}
\arguments{
\item{perm_results}{object created from perm_dist function. Contains the
distribution of maxT and STCS AND the matrix obtained from applying a
hypothesis test at each grid cell.}

\item{alpha_local}{significance threshold at the grid cell level.}

\item{alpha_global}{desired overall significance level}

\item{data_dim}{dimesions of original data. Used for calculating df for
t distributed test statistics, ignored if the test statistic is normal}
}
\description{
Threshold  data
Thresholds the data according to the corresponding multiple testing
correction. This function is used internally.
}
