% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boferroni_and_related_corrections.R
\name{walker_fn}
\alias{walker_fn}
\title{Walker min-p correction}
\usage{
walker_fn(pval, alpha = 0.05)
}
\arguments{
\item{pval}{matrix of p-values}

\item{alpha}{desired alpha level for the correction.}
}
\description{
For given matrix containing p.values, returns a TRUE/FALSE vector where TRUE
is significant using the correction.
}
