% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFree.R, R/allSeries.R, R/allGrid.R,
%   R/allStrata.R
\name{allUtils}
\alias{allFree}
\alias{allSeries}
\alias{allGrid}
\alias{allStrata}
\title{Utility functions for complete enumeration of all possible permutations}
\usage{
allFree(n, v = seq_len(n))

allSeries(n, nperms, mirror = FALSE)

allGrid(n, nperms, nr, nc, mirror, constant)

allStrata(n, control)
}
\arguments{
\item{n}{The number of observations.}

\item{v}{Numeric vector of indices. The default is \code{seq_len(n)}.}

\item{nperms}{Numeric; number of possible permutations.}

\item{mirror}{Logical; should mirroring of permutations be allowed?}

\item{nr, nc}{Integer; number of rows and columns of grid designs.}

\item{constant}{Logical; should the same permutation be used within each
block?}

\item{control}{A list describing the permutation design, as returned by
\code{\link[=how]{how()}}.}
}
\value{
A matrix of all possible permutations of \code{n} observations or of
\code{v}, given the provided options.
}
\description{
Utility functions to return the set of all permutations under different
designs. For most practical applications, such as combining designs that
permute blocks and/or observations within blocks, \code{\link[=allPerms]{allPerms()}} is required.
}
\details{
These utility functions are not designed for casual use. See \code{\link[=allPerms]{allPerms()}}
for further details.
}
\author{
Gavin Simpson
}
