% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFoo-methods.R
\name{get-methods}
\alias{get-methods}
\alias{getBlocks.default}
\alias{getWithin.default}
\alias{getStrata.default}
\alias{getType.default}
\alias{getMirror.default}
\alias{getSymmetric.default}
\alias{getConstant.default}
\alias{getPlots.default}
\alias{getRow.default}
\alias{getCol.default}
\alias{getDim.default}
\alias{getNperm.default}
\alias{getMaxperm.default}
\alias{getMinperm.default}
\alias{getComplete.default}
\alias{getMake.default}
\alias{getObserved.default}
\alias{getAllperms.default}
\alias{getControl.default}
\alias{getAllperms}
\alias{getBlocks}
\alias{getComplete}
\alias{getConstant}
\alias{getCol}
\alias{getDim}
\alias{getMake}
\alias{getMaxperm}
\alias{getMinperm}
\alias{getMirror}
\alias{getSymmetric}
\alias{getNperm}
\alias{getObserved}
\alias{getPlots}
\alias{getRow}
\alias{getStrata}
\alias{getType}
\alias{getWithin}
\alias{getControl}
\alias{getHow}
\alias{getAllperms.how}
\alias{getBlocks.how}
\alias{getCol.how}
\alias{getCol.Plots}
\alias{getCol.Within}
\alias{getComplete.how}
\alias{getConstant.how}
\alias{getConstant.Within}
\alias{getDim.how}
\alias{getDim.Plots}
\alias{getDim.Within}
\alias{getMake.how}
\alias{getMaxperm.how}
\alias{getMinperm.how}
\alias{getMirror.how}
\alias{getMirror.Plots}
\alias{getMirror.Within}
\alias{getSymmetric.how}
\alias{getNperm.how}
\alias{getSymmetric.Plots}
\alias{getObserved.how}
\alias{getPlots.how}
\alias{getSymmetric.Within}
\alias{getRow.how}
\alias{getRow.Plots}
\alias{getRow.Within}
\alias{getStrata.how}
\alias{getStrata.Plots}
\alias{getType.how}
\alias{getType.Plots}
\alias{getType.Within}
\alias{getWithin.how}
\alias{getControl.allPerms}
\title{Extractor functions to access components of a permutation design}
\usage{
getAllperms(object, ...)

getBlocks(object, ...)

getComplete(object, ...)

getConstant(object, ...)

getCol(object, ...)

getDim(object, ...)

getMake(object, ...)

getMaxperm(object, ...)

getMinperm(object, ...)

getMirror(object, ...)

getSymmetric(object, ...)

getNperm(object, ...)

getObserved(object, ...)

getPlots(object, ...)

getRow(object, ...)

getStrata(object, ...)

getType(object, ...)

getWithin(object, ...)

getControl(object, ...)

getHow(object, ...)

\method{getAllperms}{how}(object, ...)

\method{getBlocks}{how}(object, ...)

\method{getCol}{how}(object, which = c("plots", "within"), ...)

\method{getCol}{Plots}(object, ...)

\method{getCol}{Within}(object, ...)

\method{getComplete}{how}(object, ...)

\method{getConstant}{how}(object, ...)

\method{getConstant}{Within}(object, ...)

\method{getDim}{how}(object, which = c("plots", "within"), ...)

\method{getDim}{Plots}(object, ...)

\method{getDim}{Within}(object, ...)

\method{getMake}{how}(object, ...)

\method{getMaxperm}{how}(object, ...)

\method{getMinperm}{how}(object, ...)

\method{getMirror}{how}(object, which = c("plots", "within"), ...)

\method{getMirror}{Plots}(object, ...)

\method{getMirror}{Within}(object, ...)

\method{getSymmetric}{how}(object, which = c("plots", "within"), ...)

\method{getNperm}{how}(object, ...)

\method{getSymmetric}{Plots}(object, ...)

\method{getObserved}{how}(object, ...)

\method{getPlots}{how}(object, ...)

\method{getSymmetric}{Within}(object, ...)

\method{getRow}{how}(object, which = c("plots", "within"), ...)

\method{getRow}{Plots}(object, ...)

\method{getRow}{Within}(object, ...)

\method{getStrata}{how}(object, which = c("plots", "blocks"), drop = TRUE, ...)

\method{getStrata}{Plots}(object, drop = TRUE, ...)

\method{getType}{how}(object, which = c("plots", "within"), ...)

\method{getType}{Plots}(object, ...)

\method{getType}{Within}(object, ...)

\method{getWithin}{how}(object, ...)

\method{getControl}{allPerms}(object, ...)
}
\arguments{
\item{object}{An R object on which to dispatch.}

\item{...}{Arguments passed to other methods.}

\item{which}{Character; the level of restriction for which information is
required.}

\item{drop}{Logical; should unused factor levels be dropped?}
}
\value{
The contents of the corresponding component of \code{object}.
}
\description{
These functions provide abstracted access to components of permutation
designs such as those returned by \code{\link[=how]{how()}}. Using them instead of directly
indexing the underlying list allows that representation to evolve without
breaking user code.
}
\details{
\code{getHow()} is an alias for \code{getControl()}; specific \code{getControl()} methods
are useful when debugging.
}
\examples{
hh <- how()
getWithin(hh)
getNperm(hh)
}
\seealso{
\code{\link[=check]{check()}} checks a permutation design described by \code{\link[=how]{how()}}.
}
\author{
Gavin Simpson
}
\keyword{methods}
\keyword{utils}
