% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numPerms.R
\name{numPerms}
\alias{numPerms}
\title{Number of possible permutations for a given object}
\usage{
numPerms(object, control = how(), check = TRUE)
}
\arguments{
\item{object}{any object handled by \code{\link[stats:nobs]{stats::nobs()}}.}

\item{control}{a list of control values describing properties of the
permutation design, as returned by a call to \code{\link[=how]{how()}}.}

\item{check}{logical; should \code{control} be checked for problems?}
}
\value{
The (numeric) number of possible permutations of observations in
\code{object}.
}
\description{
\code{numPerms} calculates the maximum number of permutations possible under
the current permutation scheme.
}
\details{
Function \code{numPerms} returns the number of permutations for the passed
\code{object} and the selected permutation scheme. \code{object} can be one
of a data frame, matrix, an object for which a scores method exists, or a
numeric or integer vector. In the case of a numeric or integer vector, a
vector of length 1 can be used and it will be expanded to a vector of length
\code{object} (i.e., \code{1:object}) before computing the number of
permutations. As such, \code{object} can be the number of observations not
just the object containing the observations.

For \code{Plots(type = "partition")}, if the group sizes are \eqn{n_1,
\ldots, n_K}, the number of distinct assignments is \eqn{n! / \prod_k n_k!}.
With blocks, this quantity is calculated within each block and the results
are multiplied.
}
\note{
In general, mirroring \code{"series"} designs doubles the number of
permutations and mirroring \code{"grid"} designs can quadruple it (within
levels of strata if present). For grids with \code{symmetric = TRUE}, at most
three orientations are included because simultaneous row and column
mirroring is disallowed. Reflections of grid axes containing one or two
cells are equivalent to toroidal shifts and do not add distinct
permutations.

Mirroring does not double the number of series permutations when the series
contains only two observations.

For example, with 2 observations there are 2 permutations for
\code{"series"} designs:
\enumerate{
\item 1-2, and
\item 2-1.
}

If these two permutations were mirrored, we would have:
\enumerate{
\item 2-1, and
\item 1-2.
}

It is immediately clear that this is the same set of permutations without
mirroring (if one reorders the rows).
}
\examples{

## permutation design --- see ?how
ctrl <- how() ## defaults to freely exchangeable

## vector input
v <- 1:10
(obs <- nobs(v))
numPerms(v, control = ctrl)

## integer input
len <- length(v)
(obs <- nobs(len))
numPerms(len, control = ctrl)

## new design, objects are a time series
ctrl <- how(within = Within(type = "series"))
numPerms(v, control = ctrl)
## number of permutations possible drastically reduced...
## ...turn on mirroring
ctrl <- how(within = Within(type = "series", mirror = TRUE))
numPerms(v, control = ctrl)

## Try blocking --- 2 groups of 5
bl <- numPerms(v, control = how(blocks = gl(2,5)))
bl

## should be same as
pl <- numPerms(v, control = how(plots = Plots(strata = gl(2,5))))
pl
stopifnot(all.equal(bl, pl))

## Distinct assignments to groups of sizes 3 and 2
groups <- factor(c("a", "a", "a", "b", "b"))
ctrl <- how(plots = Plots(groups, type = "partition"))
numPerms(length(groups), control = ctrl) ## 10

}
\seealso{
\code{\link[=shuffle]{shuffle()}} and \code{\link[=how]{how()}}. Additional
\code{\link[stats:nobs]{stats::nobs()}} methods are provided; see \link{nobs-methods}.
}
\author{
Gavin Simpson
}
