% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFoo-methods.R
\name{set-methods}
\alias{set-methods}
\alias{setBlocks<-.default}
\alias{setBlocks<-.how}
\alias{setWithin<-.default}
\alias{setWithin<-.how}
\alias{setStrata<-.default}
\alias{setStrata<-.how}
\alias{setStrata<-.Plots}
\alias{setType<-.default}
\alias{setType<-.how}
\alias{setType<-.Plots}
\alias{setType<-.Within}
\alias{setMirror<-.default}
\alias{setMirror<-.how}
\alias{setMirror<-.Plots}
\alias{setMirror<-.Within}
\alias{setSymmetric<-.default}
\alias{setSymmetric<-.how}
\alias{setSymmetric<-.Plots}
\alias{setSymmetric<-.Within}
\alias{setConstant<-.default}
\alias{setConstant<-.how}
\alias{setConstant<-.Plots}
\alias{setConstant<-.Within}
\alias{setPlots<-.default}
\alias{setPlots<-.how}
\alias{setRow<-.default}
\alias{setRow<-.how}
\alias{setRow<-.Plots}
\alias{setRow<-.Within}
\alias{setCol<-.default}
\alias{setCol<-.how}
\alias{setCol<-.Plots}
\alias{setCol<-.Within}
\alias{setDim<-.default}
\alias{setDim<-.how}
\alias{setDim<-.Plots}
\alias{setDim<-.Within}
\alias{setNperm<-.default}
\alias{setNperm<-.how}
\alias{setAllperms<-.default}
\alias{setAllperms<-.how}
\alias{setMaxperm<-.default}
\alias{setMaxperm<-.how}
\alias{setMinperm<-.default}
\alias{setMinperm<-.how}
\alias{setComplete<-.default}
\alias{setComplete<-.how}
\alias{setMake<-.default}
\alias{setMake<-.how}
\alias{setObserved<-.default}
\alias{setObserved<-.how}
\alias{setBlocks<-}
\alias{setPlots<-}
\alias{setWithin<-}
\alias{setStrata<-}
\alias{setNperm<-}
\alias{setAllperms<-}
\alias{setMaxperm<-}
\alias{setMinperm<-}
\alias{setComplete<-}
\alias{setMake<-}
\alias{setObserved<-}
\alias{setRow<-}
\alias{setCol<-}
\alias{setDim<-}
\alias{setType<-}
\alias{setMirror<-}
\alias{setSymmetric<-}
\alias{setConstant<-}
\title{Replacement functions to set components of a permutation design}
\usage{
setBlocks(object) <- value

setPlots(object) <- value

setWithin(object) <- value

setStrata(object) <- value

setNperm(object) <- value

setAllperms(object) <- value

setMaxperm(object) <- value

setMinperm(object) <- value

setComplete(object) <- value

setMake(object) <- value

setObserved(object) <- value

setRow(object) <- value

setCol(object) <- value

setDim(object) <- value

setType(object) <- value

setMirror(object) <- value

setSymmetric(object) <- value

setConstant(object) <- value
}
\arguments{
\item{object}{An R object on which to dispatch.}

\item{value}{The replacement value or object.}
}
\value{
\code{object}, suitably modified.
}
\description{
These functions provide abstracted replacement of components in a
permutation design such as one returned by \code{\link[=how]{how()}}. They also update the
matched calls stored in the design so that \code{\link[stats:update]{stats::update()}} continues to
work correctly.
}
\details{
Use these functions instead of directly modifying the underlying list, so
code does not depend on permute's internal representation.
}
\section{Note}{

\verb{setStrata<-} has methods for objects of class \code{"how"} and \code{"Plots"}. The
former sets the \code{blocks} component of the \code{\link[=how]{how()}} object, while the latter
sets the \code{strata} component of the \code{\link[=Plots]{Plots()}} object.

\verb{setDim<-}, \verb{setRow<-}, and \verb{setCol<-} cannot be used on an object of class
\code{"how"}. Instead, extract the \code{\link[=Plots]{Plots()}} or \code{\link[=Within]{Within()}} component with
\code{\link[=getPlots]{getPlots()}} or \code{\link[=getWithin]{getWithin()}}, alter it, and replace it with \verb{setPlots<-} or
\verb{setWithin<-}.
}

\examples{
hh <- how()
getNperm(hh)
setNperm(hh) <- 999
getNperm(hh)
}
\seealso{
\code{\link[=check]{check()}} checks a design described by \code{\link[=how]{how()}}. See \link{get-methods}
for the corresponding extractor functions.
}
\author{
Gavin Simpson
}
\keyword{methods}
\keyword{utils}
